chore (refs T22614): move store of Statement to client/js
