+++ client/js/bundles/core/publicParagraphDocument.js
@@ -15,7 +15,7 @@ import DpPublicDetailNoMap from '@DpJs/components/statement/DpPublicDetailNoMap'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import { prefixClass } from 'demosplan-ui/lib'
-import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
+import publicStatement from '@DpJs/store/statement/PublicStatement'
 import TableWrapper from '@DpJs/lib/TableWrapper'
 import TocStateMemorizer from '@DemosPlanCoreBundle/lib/TocStateMemorizer'
