+++ client/js/bundles/map/publicDetail.js
@@ -23,7 +23,7 @@ import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import layers from '@DpJs/store/map/Layers'
 import Map from '@DpJs/components/map/publicdetail/Map'
 import MapTools from '@DpJs/components/map/publicdetail/controls/MapTools'
-import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
+import publicStatement from '@DpJs/store/statement/PublicStatement'
 import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
 import StatementForm from '@DpJs/lib/statement/StatementForm'
 import TableWrapper from '@DpJs/lib/TableWrapper'
