+++ client/js/bundles/statement/assessmentStatement.js
@@ -12,13 +12,13 @@
  */
 
 import AssessmentStatement from '@DpJs/lib/statement/AssessmentStatement'
-import AssessmentTableStore from '@DemosPlanStatementBundle/store/AssessmentTable'
+import AssessmentTableStore from '@DpJs/store/statement/AssessmentTable'
 import DetailView from '@DpJs/components/statement/assessmentTable/DetailView/DetailView'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import dpValidate from '@DpJs/lib/core/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import StatementStore from '@DemosPlanStatementBundle/store/Statement'
-import VoterStore from '@DemosPlanStatementBundle/store/Voter'
+import StatementStore from '@DpJs/store/statement/Statement'
+import VoterStore from '@DpJs/store/statement/Voter'
 
 const stores = {
   assessmentTable: AssessmentTableStore,
