+++ client/js/bundles/statement/listFinalGroup.js
@@ -14,7 +14,7 @@
 import DpMapModal from '@DpJs/components/statement/assessmentTable/DpMapModal'
 import DpPublicStatementList from '@DpJs/components/statement/publicStatementLists/DpPublicStatementList'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
+import publicStatement from '@DpJs/store/statement/PublicStatement'
 
 const components = {
   DpMapModal,
