+++ client/js/store/statement/PublicStatement.js
