+++ client/js/store/statement/Voter.js
